Мелкие правки 2:
- Кнопка сортировки теперь является элементом NavifationBar
- На экране коллекции теперь все элементы в ScrollView
- Реализована функция динамического обновления высоты коллекции

+++ FakeNFT/Scenes /Catalog/ViewModel/NFTCellModel.swift
@@ -12,7 +12,7 @@ struct NFTCellModel {
     let image: UIImage
     let rating: Int
     let name: String
-    let cost: Float
+    let cost: Int
     let isLike: Bool
     let inCart: Bool
